+++ src/operations/DivideOperation.kt
@@ -1,8 +1,8 @@
 package operations
 
 class DivideOperation : AbstractOperation() {
-    override fun calc(a: Int, b: Int): Int {
+    override fun calc(a: Long, b: Long): Long {
         val divideResult = a.toDouble() / b
-        return divideResult.toInt()
+        return divideResult.toLong()
     }
 }
