+++ src/operations/SubtractOperation.kt
@@ -1,7 +1,7 @@
 package operations
 
 class SubtractOperation : AbstractOperation() {
-    override fun calc(a: Int, b: Int): Int {
+    override fun calc(a: Long, b: Long): Long {
         return a - b
     }
 }
